dockerregistry: Move manifest, context, registry logic
